refactor(cli): extract commands.rs from main.rs
The 1870-line main.rs mixed CLI bootstrap, arg dispatch, per-command
business logic, the persistence layer, and the entire TUI surface
(run_tui + handle_tui_command, ~1100 lines). This first pass extracts
the per-command business logic + persistence helpers into a new
commands.rs module.

Moved out of main.rs:
- run_discover, run_scan, run_inspect, run_sweep, run_dns, run_reverse
  (the per-command 'do the work, persist to history, return result'
  functions)
- print_dns_records_text (DNS-specific text formatting)
- try_init_db, init_db (SQLite history database setup)
- db_upsert_host_safe, db_add_scan_history_safe,
  db_upsert_discovered_hosts_safe, db_upsert_sweep_hosts_safe (the
  best-effort 'log on failure' wrappers around Database calls)
- warn_nonfatal helper

main.rs is now 1674 lines (down from 1870). The remaining bulk is
the TUI run loop (run_tui at ~366 lines + handle_tui_command at ~755
lines) which lives next to the TUI code conceptually — Tier 1 #2's
tui.rs split will pull those out too.

Pub surface of commands.rs is just what main.rs / tui.rs need to
call: run_*, print_dns_records_text, try_init_db, and
db_add_scan_history_safe (which the TUI handler also uses for
recording scan history when running interactively).

Verified locally:
- cargo check -p netscli clean
- cargo clippy -p netscli --all-targets -- -D warnings clean
- cargo fmt --all -- --check clean
- cargo test -p netscli — 25 tests pass
